Rename parser to workspace
